Changes to enable the JSDO to use xmlhttprequest-cookie
When working with the xmlhttprequest-cookie library, objects created using the wrapper are not recognized as instances of XMLHttpRequest. (instanceof XMLHttpRequest).
Specified "XMLHttpRequest" as the function name so that code can check for constructor.name.

The new behavior is consistent with web browser behavior where constructor.name for an XHR object is set.

When working with FORM-based authentication from the JSDO.
The Accept header always includes "*/*" because it is listed in the defaultHeaders variable.
Code in the JSDOSession expects to be able to set the Accept header to "application/json".
Changed line 545 to copy defaultHeaders to the headers variable and then delete the "Accept" property so that it can be set by calling setRequestHeader(). If the Accept header is not specified, then code in line 337 ensures that it is set to the default: "*/*".

The new behavior is consistent with web browser behavior where the Accept header is not set by default.
See:
https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequest/setRequestHeader